DB: change Id to long type

+++ api/models/BaseEntity.cs
@@ -6,7 +6,7 @@ namespace Budgeting.Models
 {
     public abstract class BaseEntity
     {
-        public int Id { get; set; } // Or Guid if you prefer
+        public long Id { get; set; } // Or Guid if you prefer
         public DateTime CreatedAt { get; set; }
         public DateTime UpdatedAt { get; set; }
     }
